+++ packages/bench/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @clippium/bench
 
+## 1.0.4
+
+### Patch Changes
+
+- up dev deps
+
 ## 1.0.3
 
 ### Patch Changes