+++ packages/preset/default/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @clippium/preset-colored
 
+## 1.0.4
+
+### Patch Changes
+
+- up dev deps
+
+- Updated dependencies []:
+  - clippium@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes